+++ spec/unit/sandbox/file_accessor_spec.rb
@@ -117,6 +117,7 @@ module Pod
           @root + 'Resources/Base.lproj',
           @root + 'Resources/Images.xcassets',
           @root + 'Resources/Migration.xcmappingmodel',
+          @root + 'Resources/Sample.rcproject',
           @root + 'Resources/Sample.xcdatamodeld',
           @root + 'Resources/de.lproj',
           @root + 'Resources/en.lproj',
@@ -169,6 +170,7 @@ module Pod
           @root + 'Resources/en.lproj',
           @root + 'Resources/Images.xcassets',
           @root + 'Resources/Migration.xcmappingmodel',
+          @root + 'Resources/Sample.rcproject',
           @root + 'Resources/Sample.xcdatamodeld',
           @root + 'Resources/sub_dir',
         ]
@@ -184,6 +186,7 @@ module Pod
           @root + 'Resources/en.lproj',
           @root + 'Resources/Images.xcassets',
           @root + 'Resources/Migration.xcmappingmodel',
+          @root + 'Resources/Sample.rcproject',
           @root + 'Resources/Sample.xcdatamodeld',
           @root + 'Resources/sub_dir',
         ]
@@ -199,6 +202,7 @@ module Pod
           @root + 'Resources/en.lproj',
           @root + 'Resources/Images.xcassets',
           @root + 'Resources/Migration.xcmappingmodel',
+          @root + 'Resources/Sample.rcproject',
           @root + 'Resources/Sample.xcdatamodeld',
           @root + 'Resources/sub_dir',
         ]