+++ CHANGELOG.md
@@ -35,6 +35,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 * Add flag to ignore prerelease versions when reporting latest version for outdated pods.  
   [cltnschlosser](https://github.com/cltnschlosser)
   [#9916](https://github.com/CocoaPods/CocoaPods/pull/9916)
+  
+* Enable the usage of RealityComposer-projects 
+  Add RealityComposer-projects (rcproject) that are defined in the resources-spec will be compiled.
+  [Hendrik von Prince](https://github.com/parallaxe)
 
 ##### Bug Fixes